requirements.txt
# LiteLLM и прокси
litellm>=1.50.0,<2.0.0
litellm[proxy]>=1.50.0,<2.0.0

# HTTP клиент
requests>=2.31.0,<3.0.0

# Сертификаты
certifi>=2023.7.22

# Переменные окружения
python-dotenv>=1.0.0,<2.0.0

# Стандартные библиотеки (обычно не нужно указывать, но для ясности)
# importlib - встроенная библиотека Python 3.1+
